dom/base/test/test_bug622117.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=622117
-->
<head>
  <title>Test for Bug 622117</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=622117">Mozilla Bug 622117</a>
<p id="display">
  <iframe id="testframe"
          srcdoc="<a href='PASS.html' onclick='throw 1'>Click me</a>">
  </iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 622117 **/
SimpleTest.waitForExplicitFinish();

const PASSURL = "data:text/plain,PASS";

SimpleTest.waitForFocus(function() {
  $("testframe").onload = function() {
    is(this.contentDocument.documentElement.innerText, "PASS", "Should have loaded link");
    SimpleTest.finish();
  };

  var win = $("testframe").contentWindow;
  var a = win.document.getElementsByTagName("a")[0];
  synthesizeMouseAtCenter(a, {}, win);
});
</script>
</pre>
</body>
</html>
